Add isort precommit
