+++ .pre-commit-config.yaml
@@ -16,11 +16,12 @@ repos:
     - id: check-ast
     - id: fix-encoding-pragma
     - id: requirements-txt-fixer
+-   repo: https://github.com/psf/black
+    rev: stable
+    hooks:
+    - id: black
 -   repo: https://github.com/timothycrosley/isort
     rev: 4.3.21
     hooks:
     - id: isort
--   repo: https://github.com/psf/black
-    rev: 19.3b0
-    hooks:
-    - id: black
+      additional_dependencies: ["toml"]
