+++ cellrank/__init__.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
-import cellrank.tools as tl
-import cellrank.plotting as pl
-import cellrank.utils as ul
-import cellrank.datasets
-
-from cellrank.tools._constants import Lin
-from cellrank.tools._read import read
 from scanpy import settings
 
+import cellrank.tools as tl
+import cellrank.utils as ul
+import cellrank.datasets
+import cellrank.plotting as pl
+from cellrank.tools._read import read
+from cellrank.tools._constants import Lin
 
 __author__ = ", ".join(["Marius Lange", "Michal Klein", "Juan Luis Restrepo Lopez"])
 __email__ = ", ".join(["info@cellrank.org"])
