+++ cellrank/datasets/_datasets.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 import os
 from typing import Union
+
 from scanpy import read
 from anndata import AnnData
