+++ cellrank/plotting/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-from cellrank.plotting._gene_trend import gene_trends
-from cellrank.plotting._heatmap import heatmap
-from cellrank.plotting._cluster_lineages import cluster_lineage
-from cellrank.plotting._cluster_fates import cluster_fates, similarity_plot
 from cellrank.plotting._graph import graph
 from cellrank.plotting._utils import lineages, composition
+from cellrank.plotting._heatmap import heatmap
+from cellrank.plotting._gene_trend import gene_trends
+from cellrank.plotting._cluster_fates import cluster_fates, similarity_plot
+from cellrank.plotting._cluster_lineages import cluster_lineage
