+++ cellrank/tools/__init__.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
-from cellrank.tools._gene_importance import gene_importance
-from cellrank.tools._utils import partition, cyto_trace
-from cellrank.tools._transition_matrix import transition_matrix
-from cellrank.tools._exact_mc_test import exact_mc_perm_test
-from cellrank.tools.estimators import CFLARE, GPCCA
-from cellrank.tools._root_final import find_root, find_final
-from cellrank.tools._lineages import lineages
-from cellrank.tools._lineage import Lineage
-
 import cellrank.tools.kernels
+from cellrank.tools._utils import partition, cyto_trace
+from cellrank.tools._lineage import Lineage
+from cellrank.tools._lineages import lineages
+from cellrank.tools.estimators import GPCCA, CFLARE
+from cellrank.tools._root_final import find_root, find_final
+from cellrank.tools._exact_mc_test import exact_mc_perm_test
+from cellrank.tools._gene_importance import gene_importance
+from cellrank.tools._transition_matrix import transition_matrix
