+++ cellrank/tools/estimators/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
-from cellrank.tools.estimators._cflare import CFLARE
 from cellrank.tools.estimators._gppca import GPCCA
+from cellrank.tools.estimators._cflare import CFLARE
