+++ cellrank/tools/kernels/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 from cellrank.tools.kernels._kernel import (
     Kernel,
-    ConnectivityKernel,
-    VelocityKernel,
-    PalantirKernel,
     Constant,
+    PalantirKernel,
+    VelocityKernel,
+    ConnectivityKernel,
 )
