+++ cellrank/utils/models/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from cellrank.utils.models._models import Model, SKLearnModel, GamMGCVModel
+from cellrank.utils.models._models import Model, GamMGCVModel, SKLearnModel
