+++ pyproject.toml
@@ -8,4 +8,19 @@ pytest = { version = "^5.3", python = "^3.6" }
 
 [tool.black]
 line-length = 88
-target_version = ['py36']
+target-version = ['py36', 'py37', 'py38']
+
+[tool.isort]
+line_length = 88
+multi_line_output = 3
+include_trailing_comma = true
+use_parentheses = true
+known_bio="scanpy,scvelo,anndata"
+known_plot="mpl_toolkits,matplotlib,seaborn"
+known_num="numpy,pandas,networkx,scipy,sklearn,statmodels"
+sections = "FUTURE,STDLIB,THIRDPARTY,PLOT,BIO,FIRSTPARTY,LOCALFOLDER"
+no_lines_before="LOCALFOLDER"
+balanced_wrapping = true
+force_grid_wrap = 0
+length_sort = "1"
+indent = "    "
