+++ setup.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
-from setuptools import setup, find_packages
+import os
 from pathlib import Path
 
-import os
+from setuptools import setup, find_packages
 
 try:
     from cellrank import __author__, __email__, __version__
